backend/requirements.txt
google-genai
dotenv
fastapi
uvicorn
google-generativeai